zephyr: Move build_and_flash into boards/nrf5x.py
to make it more customizable.
