+++ autopts/ptsprojects/boards/reel_board.py
@@ -17,6 +17,9 @@
 supported_projects = ['zephyr']
 
 
+board_type = 'reel_board'
+
+
 def reset_cmd(iutctl):
     """Return reset command for Reel_Board DUT
